rename "npm run update-screenshots" to "npm run render-goldens", more accurate name.
