+++ packages/render-fidelity-tools/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "node --experimental-modules ./lib/workflows/test-fidelity.js --config=./test/config.json",
     "test:ci": "./scripts/ci-check-fidelity.sh",
-    "update-screenshots": "node --experimental-modules ./lib/workflows/update-screenshots.js --config=./test/config.json",
+    "render-goldens": "node --experimental-modules ./lib/workflows/render-goldens.js --config=./test/config.json",
     "build": "tsc && rollup -c",
     "prepare": "if [ ! -L './shared-assets' ]; then ln -s ../shared-assets ./shared-assets; fi",
     "clean": "rm -rf ./dist ./lib ./test/results ./test/renderers/filament/{cmgen,ktx}"
