+++ packages/render-fidelity-tools/src/workflows/render-goldens/renderer-offline.ts
