+++ packages/render-fidelity-tools/src/workflows/render-goldens/renderer-screenshot.ts
